+++ packages/app/src/assets/StackedBarIcon.js
@@ -1,11 +1,10 @@
 import React from 'react'
-import SvgIcon from '@material-ui/core/SvgIcon'
 import PropTypes from 'prop-types'
 
 const StackedBarIcon = ({
     style = { paddingRight: '8px', width: 24, height: 24 },
 }) => (
-    <SvgIcon viewBox="0,0,48,48" style={style}>
+    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0,0,48,48" style={style}>
         <g stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
             <rect fill="#63A4FF" x="2" y="4" width="17" height="10" />
             <rect
@@ -23,7 +22,7 @@ const StackedBarIcon = ({
             <rect fill="#9E9E9E" x="0" y="0" width="2" height="48" />
             <rect fill="#9E9E9E" x="0" y="46" width="48" height="2" />
         </g>
-    </SvgIcon>
+    </svg>
 )
 
 StackedBarIcon.propTypes = {