+++ packages/app/src/assets/YearOverYearColumnIcon.js
@@ -1,11 +1,10 @@
 import React from 'react'
-import SvgIcon from '@material-ui/core/SvgIcon'
 import PropTypes from 'prop-types'
 
 const YearOverYearColumnIcon = ({
     style = { paddingRight: '8px', width: 24, height: 24 },
 }) => (
-    <SvgIcon viewBox="0,0,48,48" style={style}>
+    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0,0,48,48" style={style}>
         <g stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
             <g>
                 <polygon points="0 0 48 0 48 48 0 48" />
@@ -19,7 +18,7 @@ const YearOverYearColumnIcon = ({
                 <polygon fill="#9E9E9E" points="0 46 48 46 48 48 0 48" />
             </g>
         </g>
-    </SvgIcon>
+    </svg>
 )
 
 YearOverYearColumnIcon.propTypes = {