+++ packages/app/src/components/Visualization/Visualization.js
@@ -25,7 +25,7 @@ import {
     NoDataOrDataElementGroupSetError,
     CombinationDEGSRRError,
 } from '../../modules/error'
-import LoadingMask from './LoadingMask'
+import LoadingMask from '../../widgets/LoadingMask'
 
 export class Visualization extends Component {
     constructor(props) {
@@ -82,6 +82,7 @@ export class Visualization extends Component {
                     id,
                     name: item.name || item.displayName,
                     displayName: item.displayName,
+                    dimensionItemType: item.dimensionItemType,
                 }
             })
         })